core/controllers/mgcp/src/test/java/org/mobicents/media/server/ctrl/mgcp/signal/MgcpPackagesTest.java
/*
 * JBoss, Home of Professional Open Source
 * Copyright 2011, Red Hat, Inc. and individual contributors
 * by the @authors tag. See the copyright.txt in the distribution for a
 * full listing of individual contributors.
 *
 * This is free software; you can redistribute it and/or modify it
 * under the terms of the GNU Lesser General Public License as
 * published by the Free Software Foundation; either version 2.1 of
 * the License, or (at your option) any later version.
 *
 * This software is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this software; if not, write to the Free
 * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */

/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */

package org.mobicents.media.server.ctrl.mgcp.signal;

import jain.protocol.ip.mgcp.message.parms.EventName;
import jain.protocol.ip.mgcp.pkg.MgcpEvent;
import jain.protocol.ip.mgcp.pkg.PackageName;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author kulikov
 */
public class MgcpPackagesTest {
    private MgcpPackages mgcpPackages = new MgcpPackages();
    
    public MgcpPackagesTest() {
    }

    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @Before
    public void setUp() {
    }

    @After
    public void tearDown() {
    }

    /**
     * Test of load method, of class MgcpPackages.
     */
    @Test
    public void testLoad() throws Exception {
        SignalExecutor pkg = mgcpPackages.load("D");
        assertTrue("Package was not loaded", pkg != null);
        
        EventName eventName = new EventName(PackageName.Dtmf, MgcpEvent.dtmf0);
        System.out.println(eventName);
        
        Signal signal = pkg.getSignal(new EventName(PackageName.Dtmf, MgcpEvent.dtmf0));
        assertTrue("Signal is null", signal != null);
    }

    @Test
    public void testLoadAll() throws Exception {
        //just should not be exceptions
        mgcpPackages.loadAll();
    }
    
}